Adds timer, simplifies TimerTask, and does a bit of folder renaming (#24)
Main addition here is the Timer, which will end up being the main object used to schedule TimerTasks in client code. A Timer is a collection of related TimerTasks that operate using the same TimeTrackingStategy. Each TimerTask may have it's own periodicity, start, and end times.

Design takes concurrency and async programming into account as well as the idea of cancelling tasks, although the latter is not currently implemented.

Also, note the internal Update and Initialize. Timers are never meant to be instantiated or updated by client code, but rather monitored / interacted with after being created by a TimerFactory. Having these be internal allows for different implementations without exposing to client code.